Adding mutex to functions that use ForEachMaster method

+++ temporal/keyvalue/keyvalue_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"os"
 	"strings"
 	"testing"
 	"time"
@@ -676,6 +677,12 @@ func TestKeyValue_DeleteScanMatch(t *testing.T) {
 			pattern:     "",
 			expectedErr: nil,
 		},
+		{
+			name:            "closed_connection",
+			pattern:         "key*",
+			expectedDeleted: 0,
+			expectedErr:     temperr.ClosedConnection,
+		},
 	}
 
 	for _, connector := range connectors {
@@ -688,7 +695,13 @@ func TestKeyValue_DeleteScanMatch(t *testing.T) {
 
 				flusher, err := flusher.NewFlusher(connector)
 				assert.Nil(t, err)
-				defer assert.Nil(t, flusher.FlushAll(ctx))
+
+				if errors.Is(tc.expectedErr, temperr.ClosedConnection) {
+					err := connector.Disconnect(ctx)
+					assert.Nil(t, err)
+				} else {
+					defer assert.Nil(t, flusher.FlushAll(ctx))
+				}
 
 				if tc.setup != nil {
 					tc.setup(kv)
@@ -755,6 +768,12 @@ func TestKeyValue_Keys(t *testing.T) {
 			expectedKeys: []string{"test2", "test3"}, // SCAN will iterate over all the keys if no pattern provided.
 			expectedErr:  nil,
 		},
+		{
+			name:         "closed_connection",
+			pattern:      "key*",
+			expectedKeys: []string{},
+			expectedErr:  temperr.ClosedConnection,
+		},
 	}
 
 	for _, connector := range connectors {
@@ -767,7 +786,13 @@ func TestKeyValue_Keys(t *testing.T) {
 
 				flusher, err := flusher.NewFlusher(connector)
 				assert.Nil(t, err)
-				defer assert.Nil(t, flusher.FlushAll(ctx))
+
+				if errors.Is(tc.expectedErr, temperr.ClosedConnection) {
+					err := connector.Disconnect(ctx)
+					assert.Nil(t, err)
+				} else {
+					defer assert.Nil(t, flusher.FlushAll(ctx))
+				}
 
 				if tc.setup != nil {
 					tc.setup(kv)
@@ -929,6 +954,7 @@ func TestKeyValue_GetKeysAndValuesWithFilter(t *testing.T) {
 
 				flusher, err := flusher.NewFlusher(connector)
 				assert.Nil(t, err)
+
 				defer assert.Nil(t, flusher.FlushAll(ctx))
 
 				if tc.setup != nil {
@@ -951,11 +977,13 @@ func TestKeyValue_GetKeysWithOpts(t *testing.T) {
 		name                string
 		setup               func(model.KeyValue)
 		searchStr           string
-		cursor              uint64
-		count               int
+		cursor              map[string]uint64
+		count               int64
 		expectedKeysCheck   func([]string) bool
+		expectedCursorCheck func(cursorMap map[string]uint64) bool
+		continueScanCheck   bool
 		expectedErr         error
-		expectedCursorCheck func(uint64) bool
+		onlyCluster         bool // Only run certain checks on cluster connectors
 	}{
 		{
 			name: "valid_search",
@@ -965,29 +993,49 @@ func TestKeyValue_GetKeysWithOpts(t *testing.T) {
 				assert.NoError(t, redis.Set(ctx, "key1", "value1", 0))
 			},
 			searchStr: "key*",
-			cursor:    0,
+			cursor:    nil,
 			count:     10,
 			expectedKeysCheck: func(s []string) bool {
 				return len(s) == 2 && (s[0] == "key1" || s[0] == "key2") && (s[1] == "key1" || s[1] == "key2")
 			},
-			expectedErr: nil,
-			expectedCursorCheck: func(c uint64) bool {
-				return c == 0
+			expectedCursorCheck: func(cursorMap map[string]uint64) bool {
+				if len(cursorMap) == 0 {
+					return false
+				}
+
+				for _, c := range cursorMap {
+					if c != 0 {
+						return false
+					}
+				}
+
+				return true
 			},
+			expectedErr: nil,
 		},
 		{
 			name:      "empty_search",
 			setup:     nil,
 			searchStr: "",
-			cursor:    0,
+			cursor:    nil,
 			count:     10,
 			expectedKeysCheck: func(s []string) bool {
 				return len(s) == 0
 			},
-			expectedErr: nil,
-			expectedCursorCheck: func(c uint64) bool {
-				return c == 0
+			expectedCursorCheck: func(cursorMap map[string]uint64) bool {
+				if len(cursorMap) == 0 {
+					return false
+				}
+
+				for _, c := range cursorMap {
+					if c != 0 {
+						return false
+					}
+				}
+
+				return true
 			},
+			expectedErr: nil,
 		},
 		{
 			name: "specific_pattern_search",
@@ -997,31 +1045,51 @@ func TestKeyValue_GetKeysWithOpts(t *testing.T) {
 				assert.NoError(t, kv.Set(ctx, "specific2", "value2", 0))
 			},
 			searchStr: "specific*",
-			cursor:    0,
+			cursor:    nil,
 			count:     10,
 			expectedKeysCheck: func(s []string) bool {
 				return len(s) == 2 &&
 					(s[0] == "specific1" || s[0] == "specific2") &&
 					(s[1] == "specific1" || s[1] == "specific2")
 			},
-			expectedErr: nil,
-			expectedCursorCheck: func(c uint64) bool {
-				return c == 0
+			expectedCursorCheck: func(cursorMap map[string]uint64) bool {
+				if len(cursorMap) == 0 {
+					return false
+				}
+
+				for _, c := range cursorMap {
+					if c != 0 {
+						return false
+					}
+				}
+
+				return true
 			},
+			expectedErr: nil,
 		},
 		{
 			name:      "non_matching_pattern",
 			setup:     nil,
 			searchStr: "nomatch*",
-			cursor:    0,
+			cursor:    nil,
 			count:     10,
 			expectedKeysCheck: func(s []string) bool {
 				return len(s) == 0
 			},
-			expectedErr: nil,
-			expectedCursorCheck: func(c uint64) bool {
-				return c == 0
+			expectedCursorCheck: func(cursorMap map[string]uint64) bool {
+				if len(cursorMap) == 0 {
+					return false
+				}
+
+				for _, c := range cursorMap {
+					if c != 0 {
+						return false
+					}
+				}
+
+				return true
 			},
+			expectedErr: nil,
 		},
 		{
 			name: "paginated_search",
@@ -1032,7 +1100,7 @@ func TestKeyValue_GetKeysWithOpts(t *testing.T) {
 				}
 			},
 			searchStr: "pagekey*",
-			cursor:    0,
+			cursor:    nil,
 			count:     5, // Count is 1 but Redis SCAN does not guarantee that it will return 1 keys
 			expectedKeysCheck: func(s []string) bool {
 				if len(s) == 0 {
@@ -1054,11 +1122,21 @@ func TestKeyValue_GetKeysWithOpts(t *testing.T) {
 
 				return true
 			},
-			expectedErr: nil,
-			expectedCursorCheck: func(c uint64) bool {
-				// Cursor should be different than 0 as there are more keys to be fetched
-				return c != 0
+			expectedCursorCheck: func(cursorMap map[string]uint64) bool {
+				if len(cursorMap) == 0 {
+					return false
+				}
+
+				for _, c := range cursorMap {
+					if c == 0 {
+						return false
+					}
+				}
+
+				return true
 			},
+			continueScanCheck: true,
+			expectedErr:       nil,
 		},
 		{
 			name: "count_higher_than_actual_keys",
@@ -1069,28 +1147,78 @@ func TestKeyValue_GetKeysWithOpts(t *testing.T) {
 				}
 			},
 			searchStr: "cursorkey*",
-			cursor:    0,
-			count:     50,
+			cursor:    nil,
+			count:     100,
 			expectedKeysCheck: func(s []string) bool {
 				return len(s) == 15
 			},
-			expectedErr: nil,
-			expectedCursorCheck: func(c uint64) bool {
-				return c == 0
+			expectedCursorCheck: func(cursorMap map[string]uint64) bool {
+				if len(cursorMap) == 0 {
+					return false
+				}
+
+				for _, c := range cursorMap {
+					if c != 0 {
+						return false
+					}
+				}
+
+				return true
 			},
+			expectedErr: nil,
+		},
+		{
+			name: "mixed_cursors_in_cluster",
+			setup: func(kv model.KeyValue) {
+				ctx := context.Background()
+				for i := 0; i < 15; i++ {
+					assert.NoError(t, kv.Set(ctx, fmt.Sprintf("mixedkey%d", i), fmt.Sprintf("value%d", i), 0))
+				}
+			},
+			searchStr: "mixed*",
+			cursor:    nil,
+			count:     2,
+			expectedKeysCheck: func(s []string) bool {
+				return len(s) != 0
+			},
+			expectedCursorCheck: func(cursorMap map[string]uint64) bool {
+				// Ensuring some cursors are zero and others are not
+				var zeroExists, nonZeroExists bool
+				for _, c := range cursorMap {
+					if c == 0 {
+						zeroExists = true
+					} else {
+						nonZeroExists = true
+					}
+				}
+				return zeroExists && nonZeroExists
+			},
+			continueScanCheck: true,
+			expectedErr:       nil,
+			onlyCluster:       true,
 		},
 		{
 			name:      "test_with_error_condition",
 			searchStr: "keys*",
-			cursor:    0,
+			cursor:    nil,
 			count:     10,
 			expectedKeysCheck: func(s []string) bool {
 				return false
 			},
-			expectedErr: temperr.ClosedConnection,
-			expectedCursorCheck: func(c uint64) bool {
-				return c == 0
+			expectedCursorCheck: func(cursorMap map[string]uint64) bool {
+				if len(cursorMap) == 0 {
+					return false
+				}
+
+				for _, c := range cursorMap {
+					if c != 0 {
+						return false
+					}
+				}
+
+				return true
 			},
+			expectedErr: temperr.ClosedConnection,
 		},
 	}
 
@@ -1103,7 +1231,10 @@ func TestKeyValue_GetKeysWithOpts(t *testing.T) {
 				} else {
 					flusher, err := flusher.NewFlusher(connector)
 					assert.Nil(t, err)
-					defer assert.Nil(t, flusher.FlushAll(ctx))
+					defer func(ctx context.Context) {
+						err := flusher.FlushAll(ctx)
+						assert.Nil(t, err)
+					}(ctx)
 				}
 				kv, err := NewKeyValue(connector)
 				assert.Nil(t, err)
@@ -1112,12 +1243,19 @@ func TestKeyValue_GetKeysWithOpts(t *testing.T) {
 					tc.setup(kv)
 				}
 
-				keys, cursor, err := kv.GetKeysWithOpts(ctx, tc.searchStr, tc.cursor, tc.count)
+				keys, newCursor, continueScan, err := kv.GetKeysWithOpts(ctx, tc.searchStr, tc.cursor, tc.count)
 				assert.Equal(t, tc.expectedErr, err)
-
+				assert.Equal(t, tc.continueScanCheck, continueScan)
 				if err == nil {
 					assert.True(t, tc.expectedKeysCheck(keys))
-					assert.True(t, tc.expectedCursorCheck(cursor))
+					if tc.onlyCluster {
+						if os.Getenv("REDIS_CLUSTER") == "true" {
+							assert.True(t, tc.expectedCursorCheck(newCursor))
+						}
+						return
+					}
+
+					assert.True(t, tc.expectedCursorCheck(newCursor))
 				}
 			})
 		}
